requirements.txt
fasttext>=0.9.1
libvoikko>=4.3
mysql-connector-python>=8.0.18
pexpect>=4.7.0
